+++ HISTORY.md
@@ -1,7 +1,10 @@
-# In Progress
+# In-Progress
 
 ## Improvements
 * Refactor display of TileDB objects in Jupyter notebooks to be more readable [#1049](https://github.com/TileDB-Inc/TileDB-Py/pull/1049)
+
+## Bug Fixes
+* `Dim.shape` correctly errors out if type is not integer or datetime [#1055](https://github.com/TileDB-Inc/TileDB-Py/pull/1055)
 
 # TileDB-Py 0.14.2 Release Notes